[android] #2631 - Removing AnnotationsOptions and MultiPointOptions in favor of concrete implementation
